backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
transformers>=4.41.0
torch>=2.1.0
pillow>=10.0.0
python-multipart>=0.0.9
